[Syntax] basic instruction documentation supported
