+++ VSRAD.Syntax/Constants.cs
@@ -9,7 +9,9 @@ namespace VSRAD.Syntax
          */
         internal const string PackageGuid = "bb4a9205-af03-4a7e-8d30-7a8649cc70a4";
         internal const string FunctionListToolWindowPaneGuid = "7220fd29-7d89-42ae-a15c-c4fc6889b54f";
+        internal const string NavigationListToolWindowPaneGuid = "c518eee2-289b-47cf-a877-d48a55f13f9b";
         internal const string FunctionListCommandSetGuid = "a1d46795-2324-4b58-9f8c-aa69414c3e9e";
+        internal const string NavigationListCommandSetGuid = "0c7aa63d-bc2a-4d0e-83c2-80bbb30a5ceb";
 
         /*
          * Command ID definition
@@ -20,6 +22,7 @@ namespace VSRAD.Syntax
         internal const int FunctionListMenu = 0x1000;
         internal const int FunctionListGroup = 0x1100;
         internal const int ShowHideLineNumberCommandId = 0x102;
+        internal const int NavigationListCommandId = 0x0203;
 
         /*
          * File extensions definition
@@ -30,6 +33,7 @@ namespace VSRAD.Syntax
         internal const string FileExtensionS = ".s";
         internal const string FileExtensionAsm1 = ".gas";
         internal const string FileExtensionAsm2 = ".asm2";
+        internal const string FileExtensionAsmDoc = ".radasm";
         internal static readonly List<string> DefaultFileExtensionAsm1 = new List<string>()
         {
             FileExtensionAsm1,
@@ -47,6 +51,7 @@ namespace VSRAD.Syntax
         internal const string RadeonAsmSyntaxBaseContentType = "code";
         internal const string RadeonAsmSyntaxContentType = "RadeonAsmSyntax";
         internal const string RadeonAsm2SyntaxContentType = "RadeonAsm2Syntax";
+        internal const string RadeonAsmDocumentationContentType = "RadeonAsmDocumentation";
 
         /*
          * RadeonAsmSyntax options definition
