+++ VSRAD.Syntax/IntelliSense/IntellisenseTokenDescription.cs
@@ -168,14 +168,14 @@ namespace VSRAD.Syntax.IntelliSense
 
         private static bool GetDescriptionFromComment(DocumentAnalysis documentAnalysis, ITextSnapshot version, IEnumerable<TrackingToken> tokens, out string description)
         {
-            var commentTokens = tokens.Where(t => t.Type == documentAnalysis.LINE_COMMENT || t.Type == documentAnalysis.LINE_COMMENT);
+            var commentTokens = tokens.Where(t => t.Type == documentAnalysis.LINE_COMMENT || t.Type == documentAnalysis.BLOCK_COMMENT);
 
             if (commentTokens.Any())
             {
                 description = commentTokens
                     .First()
                     .GetText(version)
-                    .Trim(new char[] { '/', '*', ' ' });
+                    .Trim(new char[] { '/', '*', ' ', '\r', '\n' });
                 return true;
             }
             else
