+++ VSRAD.Syntax/IntelliSense/Navigation/NavigationList/NavigationListCommand.cs
@@ -1,0 +1,47 @@
+﻿using System;
+using System.ComponentModel.Design;
+using Microsoft.VisualStudio.Shell;
+using Task = System.Threading.Tasks.Task;
+
+namespace VSRAD.Syntax.IntelliSense.Navigation.NavigationList
+{
+    internal sealed class NavigationListCommand
+    {
+        public static NavigationListCommand Instance { get; private set; }
+
+        private static readonly Guid CommandSet = new Guid(Constants.NavigationListCommandSetGuid);
+        private readonly AsyncPackage package;
+
+        private NavigationListCommand(AsyncPackage package, OleMenuCommandService commandService)
+        {
+            this.package = package ?? throw new ArgumentNullException(nameof(package));
+            commandService = commandService ?? throw new ArgumentNullException(nameof(commandService));
+
+            var menuCommandID = new CommandID(CommandSet, Constants.NavigationListCommandId);
+            var menuItem = new MenuCommand(Execute, menuCommandID);
+            commandService.AddCommand(menuItem);
+        }
+
+        public static async Task InitializeAsync(AsyncPackage package)
+        {
+            // Switch to the main thread - the call to AddCommand in NavigationListCommand's constructor requires
+            // the UI thread.
+            await ThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync(package.DisposalToken);
+
+            var commandService = await package.GetServiceAsync((typeof(IMenuCommandService))) as OleMenuCommandService;
+            Instance = new NavigationListCommand(package, commandService);
+        }
+
+        private void Execute(object sender, EventArgs e)
+        {
+            package.JoinableTaskFactory.RunAsync(async delegate
+            {
+                var window = await package.ShowToolWindowAsync(typeof(NavigationList), 0, true, package.DisposalToken);
+                if ((null == window) || (null == window.Frame))
+                {
+                    throw new NotSupportedException("Cannot create tool window");
+                }
+            });
+        }
+    }
+}
