+++ VSRAD.Syntax/IntelliSense/QuickInfo/QuickInfoSource.cs
@@ -10,13 +10,13 @@ namespace VSRAD.Syntax.IntelliSense.QuickInfo
 {
     internal class QuickInfoSource : IAsyncQuickInfoSource
     {
-        private readonly NavigationTokenService _navigationService;
+        private readonly INavigationTokenService _navigationService;
         private readonly ITextBuffer _textBuffer;
         private readonly DocumentAnalysis _documentAnalysis;
 
         public QuickInfoSource(ITextBuffer textBuffer, 
             DocumentAnalysis documentAnalysis,
-            NavigationTokenService navigationService)
+            INavigationTokenService navigationService)
         {
             _textBuffer = textBuffer;
             _documentAnalysis = documentAnalysis;
@@ -31,9 +31,10 @@ namespace VSRAD.Syntax.IntelliSense.QuickInfo
 
             var extent = triggerPoint.Value.GetExtent();
 
-            var navigationToken = _navigationService.GetNaviationItem(extent).AnalysisToken;
-            if (navigationToken != null)
+            var navigationTokens = _navigationService.GetNaviationItem(extent);
+            if (navigationTokens.Count == 1)
             {
+                var navigationToken = navigationTokens[0].AnalysisToken;
                 var dataElement = IntellisenseTokenDescription.GetColorizedTokenDescription(_documentAnalysis, navigationToken);
                 if (dataElement == null)
                     return Task.FromResult<QuickInfoItem>(null);
