+++ VSRAD.Syntax/Options/InstructionListManager.cs
@@ -1,30 +1,43 @@
-﻿using System;
+﻿using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Text;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using System.IO;
 using System.Linq;
-using System.Threading.Tasks;
 using VSRAD.Syntax.Helpers;
+using VSRAD.Syntax.IntelliSense.Navigation;
+using VSRAD.Syntax.Parser;
+using Task = System.Threading.Tasks.Task;
+using Microsoft.VisualStudio.Utilities;
 
 namespace VSRAD.Syntax.Options
 {
     [Export(typeof(InstructionListManager))]
     internal sealed class InstructionListManager
     {
-        public delegate void InstructionsUpdateDelegate(IReadOnlyList<string> instructions);
+        public delegate void InstructionsUpdateDelegate(IReadOnlyDictionary<string, List<NavigationToken>> instructions);
         public event InstructionsUpdateDelegate InstructionUpdated;
 
-        public List<string> InstructionList { get; }
+        private readonly RadeonServiceProvider _serviceProvider;
+        private readonly IContentType _contentType;
+        private readonly DocumentAnalysisProvoder _documentAnalysisProvoder;
+
+        public Dictionary<string, List<NavigationToken>> InstructionList { get; }
 
         [ImportingConstructor]
-        public InstructionListManager(OptionsProvider optionsEventProvider)
+        public InstructionListManager(OptionsProvider optionsEventProvider, RadeonServiceProvider serviceProvider)
         {
-            InstructionList = new List<string>();
+            _serviceProvider = serviceProvider;
+            _contentType = _serviceProvider.ContentTypeRegistryService.GetContentType(Constants.RadeonAsmDocumentationContentType);
+
+            _documentAnalysisProvoder = new DocumentAnalysisProvoder(this);
+            InstructionList = new Dictionary<string, List<NavigationToken>>();
             optionsEventProvider.OptionsUpdated += InstructionPathsUpdated;
         }
 
         private void InstructionPathsUpdated(OptionsProvider provider) =>
-            Microsoft.VisualStudio.Shell.ThreadHelper.JoinableTaskFactory.RunAsync(() => LoadInstructionsFromDirectoriesAsync(provider.InstructionsPaths));
+            ThreadHelper.JoinableTaskFactory.RunAsync(() => LoadInstructionsFromDirectoriesAsync(provider.InstructionsPaths));
 
         public Task LoadInstructionsFromDirectoriesAsync(string dirPathsString)
         {
@@ -64,13 +77,30 @@ namespace VSRAD.Syntax.Options
         {
             try
             {
-                using (var fileStream = File.OpenRead(path))
-                using (var streamReader = new StreamReader(fileStream))
+                var buffer = CreateTextDocument(path);
+                if (buffer == null)
+                    throw new InvalidDataException($"Cannot create ITextBuffer for the {path}");
+
+                var documentAnalysis = _documentAnalysisProvoder.CreateDocumentAnalysis(buffer);
+                if (documentAnalysis.LastParserResult.Count > 0)
                 {
-                    string line;
-                    while ((line = streamReader.ReadLine()) != null)
+                    var instructions = documentAnalysis
+                        .LastParserResult[0]
+                        .Tokens
+                        .Where(t => t.Type == Parser.Tokens.RadAsmTokenType.Instruction);
+
+                    var version = documentAnalysis.CurrentSnapshot;
+                    foreach (var instruction in instructions)
                     {
-                        InstructionList.Add(line.Trim());
+                        var text = instruction.TrackingToken.GetText(version);
+                        if (InstructionList.TryGetValue(text, out var navigationTokens))
+                        {
+                            navigationTokens.Add(new NavigationToken(instruction, version));
+                        }
+                        else
+                        {
+                            InstructionList.Add(text, new List<NavigationToken>() { new NavigationToken(instruction, version) });
+                        }
                     }
                 }
             }
@@ -78,6 +108,12 @@ namespace VSRAD.Syntax.Options
             {
                 Error.ShowError(e, "instrunction file paths");
             }
+        }
+
+        private ITextBuffer CreateTextDocument(string name)
+        {
+            var document = _serviceProvider.TextDocumentFactoryService.CreateAndLoadTextDocument(name, _contentType);
+            return document.TextBuffer;
         }
     }
 }
